requirements.txt
Flask==2.1.*
Flask-Cors==3.0.8
gunicorn==19.7.1
earthengine-api>=0.1.232
flasgger==0.8.1
oauth2client==4.1.2
google-cloud-firestore==1.4.0
six==1.16.0
sqlalchemy
psycopg2
pyopenssl
pyyaml==5.4.1